Events without values must be interpreted as empty strings per spec
According to the EventSource spec (see section 7: Interpreting an event stream), a line that has a field, but no value, must be interpreted as having an empty string as a value.

"Otherwise, the string is not empty but does not contain a U+003A COLON character (:)
Process the field using the steps described below, using the whole line as the field name, and the empty string as the field value."

Additionally, see the examples provided in the spec:
"The following stream fires two events:

data

data
data

data:
The first block fires events with the data set to the empty string, as would the last block if it was followed by a blank line. The middle block fires an event with the data set to a single newline character. The last block is discarded because it is not followed by a blank line."